+++ src/main/java/com/online/exams/system/core/model/PaperCondition.java
@@ -497,52 +497,52 @@ public class PaperCondition {
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsEqualTo(Integer value) {
+        public Criteria andTotalPointsEqualTo(Double value) {
             addCriterion("total_points =", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsNotEqualTo(Integer value) {
+        public Criteria andTotalPointsNotEqualTo(Double value) {
             addCriterion("total_points <>", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsGreaterThan(Integer value) {
+        public Criteria andTotalPointsGreaterThan(Double value) {
             addCriterion("total_points >", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsGreaterThanOrEqualTo(Integer value) {
+        public Criteria andTotalPointsGreaterThanOrEqualTo(Double value) {
             addCriterion("total_points >=", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsLessThan(Integer value) {
+        public Criteria andTotalPointsLessThan(Double value) {
             addCriterion("total_points <", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsLessThanOrEqualTo(Integer value) {
+        public Criteria andTotalPointsLessThanOrEqualTo(Double value) {
             addCriterion("total_points <=", value, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsIn(List<Integer> values) {
+        public Criteria andTotalPointsIn(List<Double> values) {
             addCriterion("total_points in", values, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsNotIn(List<Integer> values) {
+        public Criteria andTotalPointsNotIn(List<Double> values) {
             addCriterion("total_points not in", values, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsBetween(Integer value1, Integer value2) {
+        public Criteria andTotalPointsBetween(Double value1, Double value2) {
             addCriterion("total_points between", value1, value2, "totalPoints");
             return (Criteria) this;
         }
 
-        public Criteria andTotalPointsNotBetween(Integer value1, Integer value2) {
+        public Criteria andTotalPointsNotBetween(Double value1, Double value2) {
             addCriterion("total_points not between", value1, value2, "totalPoints");
             return (Criteria) this;
         }
@@ -557,52 +557,52 @@ public class PaperCondition {
             return (Criteria) this;
         }
 
-        public Criteria andScoreEqualTo(Integer value) {
+        public Criteria andScoreEqualTo(Double value) {
             addCriterion("score =", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotEqualTo(Integer value) {
+        public Criteria andScoreNotEqualTo(Double value) {
             addCriterion("score <>", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreGreaterThan(Integer value) {
+        public Criteria andScoreGreaterThan(Double value) {
             addCriterion("score >", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreGreaterThanOrEqualTo(Integer value) {
+        public Criteria andScoreGreaterThanOrEqualTo(Double value) {
             addCriterion("score >=", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreLessThan(Integer value) {
+        public Criteria andScoreLessThan(Double value) {
             addCriterion("score <", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreLessThanOrEqualTo(Integer value) {
+        public Criteria andScoreLessThanOrEqualTo(Double value) {
             addCriterion("score <=", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreIn(List<Integer> values) {
+        public Criteria andScoreIn(List<Double> values) {
             addCriterion("score in", values, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotIn(List<Integer> values) {
+        public Criteria andScoreNotIn(List<Double> values) {
             addCriterion("score not in", values, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreBetween(Integer value1, Integer value2) {
+        public Criteria andScoreBetween(Double value1, Double value2) {
             addCriterion("score between", value1, value2, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotBetween(Integer value1, Integer value2) {
+        public Criteria andScoreNotBetween(Double value1, Double value2) {
             addCriterion("score not between", value1, value2, "score");
             return (Criteria) this;
         }
